config/requirements.txt
# Blockchain and Web3
web3==6.15.1
eth-account==0.11.0
eth-utils==3.0.0

# GraphQL and API
gql==3.5.0
aiohttp==3.9.3
requests==2.31.0

# Data processing and analysis
pandas==2.1.4
numpy==1.24.3
networkx==3.2.1
scipy==1.11.4

# Graph databases
neo4j==5.15.0
py2neo==2021.2.3

# Machine Learning
scikit-learn==1.3.2
torch==2.1.2
torch-geometric==2.4.0
xgboost==2.0.3

# Data storage
pymongo==4.6.1
redis==5.0.1
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
loguru==0.7.2

# Development
pytest==7.4.3
black==23.12.1
flake8==6.1.0
